feat: Recurring UX update
- We no longer make the modal appear each time
a purpose is modified.
- Recurrence becomes a sub-part of the purpose option.
- If the trip has no purpose,
  selecting the new purpose activates recurrence.
- If the trip already has a purpose,
selecting the new purpose does not activate recurrence.
