--- src/components/EditDialogs/RecurrenceEditDialog/index.js
@@ -1,3 +1,0 @@
-import RecurrenceEditDialog from './RecurrenceEditDialog'
-
-export default RecurrenceEditDialog
