+++ src/locales/fr.json
@@ -98,15 +98,7 @@
   "purpose": "Motif du déplacement",
   "recurring": {
     "title": "Récurrence",
-    "recurringTrip": "Déplacement habituel",
-    "occasionalTrip": "Déplacement occasionnel",
-    "confirmDialog": {
-      "title": "Attribuer ce motif aux déplacements similaires ?",
-      "content": "En répondant oui, ce déplacement sera défini comme habituel et servira de référence pour automatiser la qualification de vos déplacement similaires.",
-      "confirm": "Oui",
-      "decline": "Non"
-    },
-    "tooltip": "Si ce déplacement est habituel, les déplacements similaires seront automatiquement référencés avec le même motif."
+    "text": "Appliquer automatiquement ce motif aux déplacements similaires"
   },
   "analysis": {
     "mode": "Émissions de CO2",
